requirements.txt
python-pptx>=1.0.0
Pillow>=12.0.0
PyYAML>=6.0
mcp>=1.0.0
openpyxl>=3.1.0
lxml>=5.0.0
